do not use alertrix.models.DirectMessage when checking if this user can request a widget
Widgets can now only be created from the direct message between a user and an application services main account. If users request the creation of a widget from another room, we respond with a message linking to the correct room. If that does not exist yet, we link to the main user