feat: forward alarms to the recipients
alerts are sent to the application service controlled account of a company and include a list of room ids, which correspond to the room ids of units. The account lists the members of these rooms and forwards the alert to them. Forwarding uses the direct message between the user and the application service account for now. This should become a separate alarm channel room later, to keep the direct message for configuration purposes only